# docs/swagger.yaml
basePath: /
definitions:
  handler.CreateSubscriptionRequest:
    properties:
      end_date:
        example: 12-2025
        type: string
      price:
        example: 400
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        example: error
        type: string
    type: object
  handler.ListSubscriptionsResponse:
    description: Список подписок с пагинацией
    properties:
      subscriptions:
        items:
          $ref: '#/definitions/handler.SubscriptionResponse'
        type: array
    type: object
  handler.SubscriptionResponse:
    properties:
      created_at:
        example: "2025-01-15T10:30:00Z"
        type: string
      end_date:
        example: 12-2025
        type: string
      id:
        example: 1
        type: integer
      price:
        example: 400
        type: integer
      service_name:
        example: Yandex Plus
        type: string
      start_date:
        example: 07-2025
        type: string
      user_id:
        example: 60601fee-2bf1-4721-ae6f-7636e79a0cba
        type: string
    type: object
  handler.TotalCostResponse:
    properties:
      count:
        example: 3
        type: integer
      total_cost:
        example: 1200
        type: integer
    type: object
  handler.UpdateSubscriptionRequest:
    properties:
      end_date:
        example: 12-2025
        type: string
      price:
        example: 800
        type: integer
      service_name:
        example: Netflix
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для управления подписками пользователей
  title: Subscription API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Возвращает список всех подписок с пагинацией
      parameters:
      - description: Лимит (по умолчанию 10, макс 100)
        in: query
        name: limit
        type: integer
      - description: Смещение (по умолчанию 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListSubscriptionsResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создаёт новую подписку для пользователя
      parameters:
      - description: Данные подписки
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Удаляет подписку по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: Подписка удалена
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      description: Возвращает подписку по её идентификатору
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Получить подписку
      tags:
      - subscriptions
    patch:
      consumes:
      - application/json
      description: Частично обновляет подписку по ID. Все поля опциональны.
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: Поля для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.SubscriptionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/cost:
    get:
      description: Рассчитывает суммарную стоимость подписок за период с опциональной
        фильтрацией по пользователю и сервису
      parameters:
      - description: Начало периода (MM-YYYY)
        example: 01-2024
        in: query
        name: start_period
        required: true
        type: string
      - description: Конец периода (MM-YYYY)
        example: 12-2024
        in: query
        name: end_period
        required: true
        type: string
      - description: UUID пользователя
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.TotalCostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Рассчитать стоимость
      tags:
      - subscriptions
  /users/{user_id}/subscriptions:
    get:
      description: Возвращает список подписок конкретного пользователя
      parameters:
      - description: UUID пользователя
        in: path
        name: user_id
        required: true
        type: string
      - description: Лимит (по умолчанию 10, макс 100)
        in: query
        name: limit
        type: integer
      - description: Смещение (по умолчанию 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ListSubscriptionsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Подписки пользователя
      tags:
      - users
swagger: "2.0"